requirement.txt
pycryptodome==3.21.0
colorama==0.4.6
tk==0.1.0